BACKEND-15: move old code; add dummy test
