+++ backend/lib/game-core/pom.xml
@@ -11,6 +11,15 @@
     <version>0.0.1-SNAPSHOT</version>
     <name>lib-game-core</name>
     <description>The New Backend game-core library</description>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+
     <build>
         <sourceDirectory>${project.basedir}/src/main/kotlin</sourceDirectory>
     </build>
