+++ backend/lib/game-core/src/test/kotlin/ru/selemilka/game/rps/RpsGame.kt
@@ -4,6 +4,8 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.flow.Flow
+import org.springframework.context.annotation.ComponentScan
+import org.springframework.context.annotation.Configuration
 import org.springframework.stereotype.Service
 import ru.selemilka.game.core.base.Game
 import ru.selemilka.game.core.base.MessageToPlayer
@@ -11,6 +13,10 @@ import ru.selemilka.game.rps.model.RpsPlayer
 import ru.selemilka.game.rps.processor.RpsRootCommand
 import ru.selemilka.game.rps.processor.RpsRootMessage
 import ru.selemilka.game.rps.processor.RpsRootProcessor
+
+@Configuration
+@ComponentScan
+class RpsGameConfiguration
 
 @Service
 class RpsGame(
