+++ backend/lib/game-core/src/test/kotlin/ru/selemilka/game/rps/old/RpsTurn.kt
@@ -1,4 +1,4 @@
-package ru.selemilka.game.rps
+package ru.selemilka.game.rps.old
 //
 //
 //sealed interface RpsPlayerCommand : RpsCommand {
