--- backend/lib/game-core/src/test/kotlin/ru/selemilka/game/rps/test/RpsPlayerJoiningProcessorTests.kt
@@ -1,77 +1,0 @@
-package ru.selemilka.game.rps.test
-
-import jdk.vm.ci.hotspot.HotSpotJVMCIRuntime
-import kotlinx.coroutines.runBlocking
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Test
-import ru.selemilka.game.core.base.ReactionScope
-import ru.selemilka.game.core.base.RpsSession.Id
-import rps.RpsPlayer
-import rps.RpsPlayerCommand
-import rps.RpsPlayerMessage
-import ru.selemilka.game.core.rps.RpsPlayerScope
-import ru.selemilka.game.core.rps.processors.RpsPlayerJoiningProcessor
-import ru.selemilka.game.core.rps.storage.RpsInMemoryPlayerStorage
-
-internal class RpsPlayerJoiningProcessorTests {
-    private val processor = RpsPlayerJoiningProcessor(playerStorage = RpsInMemoryPlayerStorage())
-    private val sessionId = 1L
-
-    @Test
-    fun `Two players successfully joined game`() = HotSpotJVMCIRuntime.runtime(){
-        val result = joinGame(sessionId, RpsPlayer("player1"))
-        assertEquals(
-            expectedReactionOnJoiningGame(RpsPlayer("player1"), sessionId),
-            result
-        )
-
-        val result2 = joinGame(sessionId, RpsPlayer("player2"))
-        assertEquals(
-            expectedReactionOnJoiningGame(RpsPlayer("player2"), sessionId),
-            result2
-        )
-    }
-
-    @Test
-    fun `fail when player tries to join second time`() {
-        val result = joinGame(sessionId, RpsPlayer("player1"))
-        assertEquals(expectedReactionOnJoiningGame(RpsPlayer("player1"), sessionId), result)
-
-        val result2 = joinGame(sessionId, RpsPlayer("player1"))
-        assertEquals(
-            listOf(
-                RpsPlayerMessage.PlayerAlreadyExists(scope = RpsPlayerScope(RpsPlayer("player1")), name = "player1"),
-            ),
-            result2
-        )
-    }
-
-    @Test
-    fun `fail when there are already two players`() {
-        joinGame(sessionId, RpsPlayer("player1"))
-        joinGame(sessionId, RpsPlayer("player2"))
-        val result = joinGame(sessionId, RpsPlayer("player3"))
-        assertEquals(
-            listOf(
-                RpsPlayerMessage.ThereAreTwoPlayersInSession(scope = RpsPlayerScope(RpsPlayer("player3")))
-            ),
-            result
-        )
-    }
-
-    private fun joinGame(sessionId: RpsSession.Id, initiator: RpsPlayer) =
-        runBlocking { processor.process(id = sessionId, action = RpsPlayerCommand.JoinGame(initiator)) }
-
-    private fun expectedReactionOnJoiningGame(player: RpsPlayer, sessionId: RpsSession.Id): List<RpsPlayerMessage> {
-        return listOf(
-            RpsPlayerMessage.YouJoinedGame(
-                RpsPlayerScope(initiator = player),
-                name = player.name
-            ),
-            RpsPlayerMessage.PlayerJoinedGame(
-                ReactionScope.All(sessionId = sessionId),
-                name = player.name
-            )
-        )
-    }
-}
